refactor(resilience): extract lockout eviction to module (file-size cap)
