+++ open-sse/services/accountFallback.ts
@@ -44,6 +44,8 @@ import {
   buildSessionQuotaFallback,
 } from "./quotaTextCooldowns.ts";
 import { parseDayGranularityResetMs, shouldPreserveQuotaSignals } from "./quotaResetParsing.ts";
+import { evictLockoutOverflow } from "./accountFallback/lockoutEviction.ts";
+export { MODEL_LOCKOUT_EVICTION_CAP } from "./accountFallback/lockoutEviction.ts";
 
 export type ProviderProfile = {
   baseCooldownMs: number;
@@ -69,7 +71,7 @@ export type ProviderProfile = {
 };
 type JsonRecord = Record<string, unknown>;
 type RateLimitReasonValue = (typeof RateLimitReason)[keyof typeof RateLimitReason];
-type ModelLockoutEntry = {
+export type ModelLockoutEntry = {
   reason: string;
   until: number;
   lockedAt: number;
@@ -77,7 +79,7 @@ type ModelLockoutEntry = {
   lastFailureAt: number;
   resetAfterMs: number;
 };
-type ModelFailureState = {
+export type ModelFailureState = {
   failureCount: number;
   lastFailureAt: number;
   resetAfterMs: number;
@@ -390,9 +392,6 @@ export async function getRuntimeProviderProfile(provider: string | null | undefi
 // ─── Per-Model Lockout Tracking ─────────────────────────────────────────────
 // In-memory map: "provider:connectionId:model" → { reason, until, lockedAt }
 const modelLockouts = new Map<string, ModelLockoutEntry>();
-// Cap prevents unbounded growth under sustained load. Entries beyond this limit
-// are evicted (oldest first, in insertion order) during the periodic cleanup.
-export const MODEL_LOCKOUT_EVICTION_CAP = 1000;
 const modelFailureState = new Map<string, ModelFailureState>();
 
 // Aliases (e.g. "cx" → "codex") must share lockout state with their canonical
@@ -480,40 +479,10 @@ function ensureCleanupTimer() {
   }
 }
 
-/** @internal exported for testing only */
+/** @internal exported for testing only (both accessors below). */
 export function evictModelLockoutOverflow(): void {
-  // Evict oldest (insertion-order) entries when cap exceeded — but NEVER a
-  // still-active (until > now) lockout: cleanupModelLockKey() already ran on
-  // every key this tick, so anything active left here is a real, in-progress
-  // cooldown, and dropping it would wrongly let routing resume to it. If the
-  // map is still over cap purely from active entries, the cap is a soft
-  // bound in that rare case rather than a correctness trade-off.
-  if (modelLockouts.size > MODEL_LOCKOUT_EVICTION_CAP) {
-    const overflow = modelLockouts.size - MODEL_LOCKOUT_EVICTION_CAP;
-    const now = Date.now();
-    // Only expired entries are eviction candidates (oldest-first, up to the
-    // overflow count) — active ones never appear in this list at all.
-    const evictableKeys = [...modelLockouts.entries()]
-      .filter(([, entry]) => entry.until <= now)
-      .slice(0, overflow)
-      .map(([key]) => key);
-    for (const key of evictableKeys) {
-      modelLockouts.delete(key);
-      modelFailureState.delete(key);
-    }
-  }
-  if (modelFailureState.size > MODEL_LOCKOUT_EVICTION_CAP) {
-    const overflow = modelFailureState.size - MODEL_LOCKOUT_EVICTION_CAP;
-    let evicted = 0;
-    for (const key of modelFailureState.keys()) {
-      if (evicted >= overflow) break;
-      if (!modelLockouts.has(key)) modelFailureState.delete(key);
-      evicted++;
-    }
-  }
+  evictLockoutOverflow(modelLockouts, modelFailureState);
 }
-
-/** @internal exported for testing only — returns modelLockouts map size */
 export function getModelLockoutSize(): number {
   return modelLockouts.size;
 }
